pyproject.toml
[project]
name = "baozhi-rag"
version = "0.1.0"
description = "面向金融保险场景的 RAG 客服系统后端基础工程。"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
  "alibabacloud-oss-v2>=1.1.3,<2.0.0",
  "email-validator>=2.3.0,<3.0.0",
  "elasticsearch>=8.15.1,<9.0.0",
  "fastapi>=0.116.1,<1.0.0",
  "openai>=1.109.1,<2.0.0",
  "pwdlib[argon2]>=0.2.1,<1.0.0",
  "pyjwt>=2.10.1,<3.0.0",
  "pymysql>=1.1.1,<2.0.0",
  "pymilvus>=2.6.0,<3.0.0",
  "pydantic-settings>=2.10.1,<3.0.0",
  "python-docx>=1.2.0,<2.0.0",
  "python-multipart>=0.0.20,<1.0.0",
  "sqlalchemy>=2.0.43,<3.0.0",
  "uvicorn[standard]>=0.35.0,<1.0.0",
]

[dependency-groups]
dev = [
  "httpx>=0.28.1,<1.0.0",
  "mypy>=1.18.2,<2.0.0",
  "pre-commit>=4.3.0,<5.0.0",
  "pytest>=8.4.2,<9.0.0",
  "pytest-cov>=7.0.0,<8.0.0",
  "ruff>=0.13.0,<1.0.0",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["src/baozhi_rag"]
sources = ["src"]

[tool.uv]
package = true

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["B", "E", "F", "I", "N", "SIM", "UP", "W"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.mypy]
python_version = "3.13"
files = ["src", "tests"]
mypy_path = "src"
namespace_packages = true
explicit_package_bases = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
strict_equality = true
pretty = true
show_error_codes = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = ["openai"]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q --cov=src/baozhi_rag --cov-report=term-missing"
testpaths = ["tests"]

[tool.coverage.report]
skip_empty = true
show_missing = true
